refactor(auth): improve proxy config reuse ability
